@@ -9,7 +9,7 @@
   "extensionDependencies": [
     "vscode.git"
   ],
-  "version": "0.3.7",
+  "version": "0.3.8",
   "license": "Apache-2.0",
   "readme": "./README.md",
   "repository": {
@@ -39,7 +39,7 @@
   },
   "engines": {
     "vscode": "^1.64.0",
-    "node": ">=10.2.0"
+    "node": ">=16.13.0"
   },
   "activationEvents": [
     "onStartupFinished",
@@ -1500,61 +1500,61 @@
     "vscode:prepublish": ""
   },
   "dependencies": {
-    "@hediet/std": "^0.6.0",
-    "@vscode/extension-telemetry": "^0.6.2",
-    "appdirs": "^1.1.0",
-    "execa": "^5.1.1",
-    "fs-extra": "^10.1.0",
-    "js-yaml": "^4.1.0",
-    "lodash.clonedeep": "^4.5.0",
-    "lodash.get": "^4.4.2",
-    "lodash.isempty": "^4.4.0",
-    "lodash.isequal": "^4.5.0",
-    "lodash.merge": "^4.6.2",
-    "lodash.omit": "^4.5.0",
-    "tree-kill": "^1.2.2",
-    "uuid": "^8.3.2"
+    "@hediet/std": "0.6.0",
+    "@vscode/extension-telemetry": "0.6.2",
+    "appdirs": "1.1.0",
+    "execa": "5.1.1",
+    "fs-extra": "10.1.0",
+    "js-yaml": "4.1.0",
+    "lodash.clonedeep": "4.5.0",
+    "lodash.get": "4.4.2",
+    "lodash.isempty": "4.4.0",
+    "lodash.isequal": "4.5.0",
+    "lodash.merge": "4.6.2",
+    "lodash.omit": "4.5.0",
+    "tree-kill": "1.2.2",
+    "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@types/chai": "^4.3.1",
-    "@types/chai-as-promised": "^7.1.4",
-    "@types/copy-webpack-plugin": "^10.1.0",
-    "@types/fs-extra": "^9.0.13",
-    "@types/jest": "^27.4.1",
-    "@types/js-yaml": "^4.0.5",
-    "@types/lodash.clonedeep": "^4.5.7",
-    "@types/lodash.get": "^4.4.7",
-    "@types/lodash.isempty": "^4.4.7",
-    "@types/lodash.isequal": "^4.5.6",
-    "@types/lodash.merge": "^4.6.7",
-    "@types/lodash.omit": "^4.5.7",
-    "@types/mocha": "^9.1.1",
-    "@types/mock-require": "^2.0.1",
-    "@types/node": "^16.11.8",
-    "@types/react-vega": "^7.0.0",
-    "@types/sinon-chai": "^3.2.6",
-    "@types/uuid": "^8.3.4",
-    "@types/vscode": "^1.64.0",
-    "@vscode/test-electron": "^2.1.5",
-    "chai": "^4.2.0",
-    "chai-as-promised": "^7.1.1",
-    "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^10.2.4",
-    "fork-ts-checker-webpack-plugin": "^7.2.11",
-    "jest": "^28.1.2",
-    "jest-environment-node": "^28.1.2",
-    "lint-staged": "^12.3.7",
-    "mocha": "^10.0.0",
-    "mock-require": "^3.0.3",
-    "shx": "^0.3.3",
-    "sinon": "^14.0.0",
-    "sinon-chai": "^3.5.0",
-    "ts-jest": "^28.0.5",
-    "ts-loader": "^9.3.1",
-    "vsce": "^2.9.2",
-    "vscode-uri": "^3.0.2",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.9.2"
+    "@types/chai": "4.3.1",
+    "@types/chai-as-promised": "7.1.5",
+    "@types/copy-webpack-plugin": "10.1.0",
+    "@types/fs-extra": "9.0.13",
+    "@types/jest": "28.1.4",
+    "@types/js-yaml": "4.0.5",
+    "@types/lodash.clonedeep": "4.5.7",
+    "@types/lodash.get": "4.4.7",
+    "@types/lodash.isempty": "4.4.7",
+    "@types/lodash.isequal": "4.5.6",
+    "@types/lodash.merge": "4.6.7",
+    "@types/lodash.omit": "4.5.7",
+    "@types/mocha": "9.1.1",
+    "@types/mock-require": "2.0.1",
+    "@types/node": "16.11.43",
+    "@types/react-vega": "7.0.0",
+    "@types/sinon-chai": "3.2.8",
+    "@types/uuid": "8.3.4",
+    "@types/vscode": "1.64.0",
+    "@vscode/test-electron": "2.1.5",
+    "chai": "4.3.6",
+    "chai-as-promised": "7.1.1",
+    "clean-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "11.0.0",
+    "fork-ts-checker-webpack-plugin": "7.2.11",
+    "jest": "28.1.2",
+    "jest-environment-node": "28.1.2",
+    "lint-staged": "13.0.3",
+    "mocha": "10.0.0",
+    "mock-require": "3.0.3",
+    "shx": "0.3.4",
+    "sinon": "14.0.0",
+    "sinon-chai": "3.7.0",
+    "ts-jest": "28.0.5",
+    "ts-loader": "9.3.1",
+    "vsce": "2.9.2",
+    "vscode-uri": "3.0.3",
+    "webpack": "5.73.0",
+    "webpack-cli": "4.10.0"
   },
   "peerDependencies": {
     "typescript": "*"
