@@ -32,40 +32,41 @@
     "svgr": "yarn workspace dvc-vscode-webview svgr"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.30.0",
-    "@typescript-eslint/parser": "^5.30.0",
-    "@vscode/codicons": "^0.0.29",
-    "eslint": "^8.18.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-prettier-standard": "^4.0.1",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-check-file": "^1.1.0",
-    "eslint-plugin-etc": "^2.0.2",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.5.3",
-    "eslint-plugin-jsx-a11y": "^6.6.0",
-    "eslint-plugin-n": "^15.2.3",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^4.2.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
-    "eslint-plugin-sonarjs": "^0.13.0",
-    "eslint-plugin-sort-keys-fix": "^1.1.2",
-    "eslint-plugin-testing-library": "^5.1.0",
-    "eslint-plugin-unicorn": "^42.0.0",
-    "husky": "^8.0.1",
-    "jest": "^28.1.2",
-    "lint-staged": "^13.0.3",
-    "npm-run-all": "^4.1.5",
-    "nyc": "^15.1.0",
-    "prettier": "^2.7.1",
-    "prettier-config-standard": "^5.0.0",
-    "ts-node": "^10.8.1",
-    "turbo": "^1.3.1",
-    "typescript": "^4.6.3"
+    "@typescript-eslint/eslint-plugin": "5.30.5",
+    "@typescript-eslint/parser": "5.30.5",
+    "@vscode/codicons": "0.0.31",
+    "eslint": "8.19.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-config-prettier-standard": "4.0.1",
+    "eslint-config-standard": "17.0.0",
+    "eslint-plugin-check-file": "1.2.0",
+    "eslint-plugin-etc": "2.0.2",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jest": "26.5.3",
+    "eslint-plugin-jsx-a11y": "6.6.0",
+    "eslint-plugin-n": "15.2.4",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-promise": "6.0.0",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-sonarjs": "0.13.0",
+    "eslint-plugin-sort-keys-fix": "1.1.2",
+    "eslint-plugin-testing-library": "5.5.1",
+    "eslint-plugin-unicorn": "43.0.0",
+    "husky": "8.0.1",
+    "jest": "28.1.2",
+    "lint-staged": "13.0.3",
+    "npm-run-all": "4.1.5",
+    "nyc": "15.1.0",
+    "prettier": "2.7.1",
+    "prettier-config-standard": "5.0.0",
+    "ts-node": "10.8.2",
+    "turbo": "1.3.1",
+    "typescript": "4.7.4"
   },
   "resolutions": {
-    "trim": "0.0.3"
+    "trim": "1.0.1",
+    "@types/react": "^18.0.14"
   }
 }
