@@ -28,56 +28,57 @@
     "classnames": "^2.2.6",
     "lodash.clonedeep": "^4.5.0",
     "lodash.merge": "^4.6.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.2",
-    "react-table": "^7.7.0",
+    "react-table": "^7.8.0",
     "react-vega": "^7.4.4",
     "react-virtualized": "^9.22.3",
     "vega": "^5.21.0",
     "vega-lite": "^5.2.0"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^6.5.8",
-    "@storybook/addons": "^6.5.8",
-    "@storybook/builder-webpack5": "^6.5.8",
-    "@storybook/manager-webpack5": "^6.5.8",
-    "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^6.5.8",
-    "@svgr/cli": "^6.0.0",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
-    "@types/classnames": "^2.2.9",
-    "@types/jest": "^27.4.1",
-    "@types/jsdom": "^16.2.6",
-    "@types/node": "^16.11.8",
-    "@types/react": "^17.0.35",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-measure": "^2.0.8",
-    "@types/react-table": "^7.7.8",
-    "@types/react-virtualized": "^9.21.21",
-    "@types/webpack": "^5.28.0",
-    "@welldone-software/why-did-you-render": "^7.0.1",
-    "chromatic": "^6.5.3",
-    "clean-webpack-plugin": "^4.0.0",
-    "css-loader": "^6.7.1",
-    "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^7.2.11",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^28.1.2",
-    "jest-canvas-mock": "^2.4.0",
-    "jest-environment-jsdom": "^28.1.2",
-    "lint-staged": "^12.3.7",
-    "raw-loader": "^4.0.2",
-    "sass": "^1.53.0",
-    "sass-loader": "^13.0.2",
-    "storybook-addon-designs": "^6.2.1",
-    "style-loader": "^3.3.1",
-    "ts-jest": "^28.0.5",
-    "ts-loader": "^9.3.1",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.9.3"
+    "@storybook/addon-essentials": "6.5.9",
+    "@storybook/addon-interactions": "6.5.9",
+    "@storybook/addons": "6.5.9",
+    "@storybook/builder-webpack5": "6.5.9",
+    "@storybook/manager-webpack5": "6.5.9",
+    "@storybook/preset-scss": "1.0.3",
+    "@storybook/react": "6.5.9",
+    "@svgr/cli": "6.2.1",
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.3.0",
+    "@types/classnames": "2.3.1",
+    "@types/jest": "28.1.4",
+    "@types/jsdom": "16.2.14",
+    "@types/node": "16.11.43",
+    "@types/react": "18.0.14",
+    "@types/react-dom": "18.0.6",
+    "@types/react-measure": "2.0.8",
+    "@types/react-table": "7.7.12",
+    "@types/react-virtualized": "9.21.21",
+    "@types/webpack": "5.28.0",
+    "@welldone-software/why-did-you-render": "7.0.1",
+    "chromatic": "6.7.0",
+    "clean-webpack-plugin": "4.0.0",
+    "css-loader": "6.7.1",
+    "file-loader": "6.2.0",
+    "fork-ts-checker-webpack-plugin": "7.2.11",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "28.1.2",
+    "jest-canvas-mock": "2.4.0",
+    "jest-environment-jsdom": "28.1.2",
+    "lint-staged": "13.0.3",
+    "raw-loader": "4.0.2",
+    "sass": "1.53.0",
+    "sass-loader": "13.0.2",
+    "storybook-addon-designs": "6.3.1",
+    "style-loader": "3.3.1",
+    "ts-jest": "28.0.5",
+    "ts-loader": "9.3.1",
+    "webpack": "5.73.0",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.9.3"
   },
   "svgr": {
     "typescript": true,
