@@ -1,9 +1,7 @@
-import * as React from 'react'
-import * as ReactDOM from 'react-dom'
+import React from 'react'
+import ReactDOM from 'react-dom/client'
 import '../shared/style.scss'
 import { App } from './components/App'
 
-const elem = document.createElement('div')
-elem.className = 'react-root'
-document.body.append(elem)
-ReactDOM.render(<App />, elem)
+const root = ReactDOM.createRoot(document.querySelector('#root') as HTMLElement)
+root.render(<App />)
