@@ -49,7 +49,7 @@ suite('Tracked Explorer Tree Test Suite', () => {
   })
 
   afterEach(function () {
-    this.timeout(5000)
+    this.timeout(6000)
     disposable.dispose()
     return closeAllEditors()
   })
