@@ -44,7 +44,7 @@ suite('Plots Test Suite', () => {
   })
 
   afterEach(function () {
-    this.timeout(5000)
+    this.timeout(6000)
     disposable.dispose()
     return closeAllEditors()
   })
