@@ -5,7 +5,9 @@ export default {
   stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   staticDirs: ['../../extension/src/test/fixtures/plotsDiff/staticImages'],
   addons: [
+    'storybook-addon-designs',
     '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
     {
       name: '@storybook/preset-scss',
       options: {
@@ -16,8 +18,7 @@ export default {
           }
         }
       }
-    },
-    'storybook-addon-designs'
+    }
   ],
   core: {
     builder: 'webpack5'
