@@ -3,7 +3,7 @@ import React from 'react'
 
 type MockButtonProps = {
   onClick: () => void
-  children: React.ReactChildren
+  children: React.ReactNode
 }
 
 export const VSCodeButton: React.FC<MockButtonProps> = ({
@@ -15,4 +15,12 @@ export const VSCodeButton: React.FC<MockButtonProps> = ({
 
 export const VSCodeDivider: React.FC = () => {
   return <hr />
+}
+
+type MockCheckboxProps = {
+  onClick: () => void
+}
+
+export const VSCodeCheckbox: React.FC<MockCheckboxProps> = ({ onClick }) => {
+  return <input type="checkbox" onClick={onClick} />
 }
