@@ -108,7 +108,7 @@ const buildDynamicColumns = (
 
       const childColumns = buildDynamicColumns(properties, path)
 
-      const column: ColumnGroup<Experiment> | TableColumn<Experiment> = {
+      return {
         Cell,
         Header,
         accessor: pathArray && buildAccessor(pathArray),
@@ -116,8 +116,7 @@ const buildDynamicColumns = (
         group: type,
         id: path,
         name: label
-      }
-      return column
+      } as ColumnGroup<Experiment> | TableColumn<Experiment>
     })
 
 const findMaxDepth = (columns: ColumnGroup<Experiment>[], depth = 1): number =>
